tests/wpt/tests/css/CSS2/normal-flow/containing-block-percent-margin-top.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/CSS22/box.html#propdef-margin-top">
<style>
  #container { overflow:hidden; background:blue; }
  #container > div { margin-top:50%; height:50px; }
</style>
<p>There should be a blue square below.</p>
<div id="container" style="width:456px;" data-expected-width="100" data-expected-height="100">
  <div></div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script>
  document.body.offsetTop;
  document.getElementById("container").style.width = "100px";
  checkLayout("#container");
</script>
